refactor to separate general cutscene from dialogue
Still not as clean as I want. Need to figure out how animating the character to match talking should work. It'd probably be best to just make the dialogue part responsible for reporting if text is getting revealed or not - and maybe when the next dialogue unit is started. With those pieces of information, the cutscene should be able to orchestrate everything else.