refactor dialogue event triggers
Now dialogue box can report when talking starts & stops as well as when individual characters are revealed - info that's useful for coordinating sound effects and animation. The dialogue box as a whole is responsible for reporting when a new dialogue unit starts, which helps the cutscene class keep track of what part of the cutscene we're currently in.

Might be a good idea to have some concept of different components 'driving' the cutscene, dictating timing, etc. probably won't go that far in the interest of actually doing something with this, though.