fix sentence pause and talk animation
Only want to do a sentence pause if there's actually punctuation at the end of the regex'd sentence.

Talk animation issue - if there's not a pause in talking between dialogue units (which is now possible because we only sentence pause on punctuation) it wouldn't transition to playing the new unit's animation. Now this is fixed (in a messy-ish way).