Dialogue Cutscene - create system for dialogue box-based cutscenes
* create basic dialogue box

Includes metadata to control pacing & played animation for character portrait.

* add text noise and dialogue character 2

* exclude temp files

* implement text skip and continue arrow

* first steps toward dialogue cutscenes

* refactor to separate general cutscene from dialogue

Still not as clean as I want. Need to figure out how animating the character to match talking should work. It'd probably be best to just make the dialogue part responsible for reporting if text is getting revealed or not - and maybe when the next dialogue unit is started. With those pieces of information, the cutscene should be able to orchestrate everything else.

* refactor dialogue event triggers

Now dialogue box can report when talking starts & stops as well as when individual characters are revealed - info that's useful for coordinating sound effects and animation. The dialogue box as a whole is responsible for reporting when a new dialogue unit starts, which helps the cutscene class keep track of what part of the cutscene we're currently in.

Might be a good idea to have some concept of different components 'driving' the cutscene, dictating timing, etc. probably won't go that far in the interest of actually doing something with this, though.

* add name tag to dialogue box

* basic character switching logic

* fix sentence pause and talk animation

Only want to do a sentence pause if there's actually punctuation at the end of the regex'd sentence.

Talk animation issue - if there's not a pause in talking between dialogue units (which is now possible because we only sentence pause on punctuation) it wouldn't transition to playing the new unit's animation. Now this is fixed (in a messy-ish way).

* cleanup & example usage

this gives some basic idea for how the new dialogue cutscene system can be used. I can't wait to see the further adventures of Gerald and Cameron.

* add bonus cutscene for humor

This also squashes some minor bug. System as a whole could use some cleanup and character expressions would be a big plus. Still not a bad result!

* improved pause behavior for mid-sentence punctuation

* separate expression and mouth animation

Also squashed a couple of bugs related to before dialogue pauses and animation decision-making. As always, some cleanup would be a good idea.

* scale talk animation speed to match reveal speed

* reorganize file structure

* rename root folder for dialogue cutscene stuff

* fix script whitespace.

* Documented & cleanup in DialogueCutscene

* Documented DialogueCharacter

* more docs, just missing DialogueUnit now

* extra dialogueCutscene docs

* finish documentation

* add missing newline

* position character portrait according to size